requirements.txt
discord.py
pdfplumber
python-dotenv